requirements.txt
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.2
fonttools==4.61.1
fsspec==2025.12.0
greenlet==3.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
librt==0.7.7
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pathspec==1.0.2
pillow==12.1.0
platformdirs==4.5.1
protobuf==6.33.2
pynndescent==0.6.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
types-requests==2.32.4.20260107
types-tqdm==4.67.0.20250809
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.9.post2
urllib3==2.6.3
webencodings==0.5.1
